requirements.txt
pandas>=1.5.0
requests>=2.28.0
black>=22.0.0
flake8>=5.0.0
pytest>=7.0.0
python-dotenv>=0.19.0
neo4j>=5.0.0